KP-4023: Back to long in event sequence
Former-commit-id: 3c734384f3136f4a5b303f9c4db24e3f73f323dc

+++ Externals/Framework/src/Ncqrs/Eventing/ServiceModel/Bus/IPublishableEvent.cs
@@ -25,7 +25,7 @@ namespace Ncqrs.Eventing.ServiceModel.Bus
         /// An sequence of events always starts with <c>1</c>. So the first event in a sequence has the <see cref="EventSequence"/> value of <c>1</c>.
         /// </remarks>
         /// <value>A number that represents the order of where this events occurred in the sequence.</value>
-        int EventSequence { get; }
+        long EventSequence { get; }
 
         /// <summary>
         /// Id of the commit this event belongs to (usually corresponds to command id).
